pyproject.toml
[tool.poetry]
name = "joytalk"
version = "1.0.0"
description = "A bot for discord to convert text to speech"
authors = ["yusukensanta <yusuke.nakayama1218@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = ">=1.5.0"
PyNaCl = "^1.4.0"
